Struct miner::MemoryPool

pub struct MemoryPool { /* fields omitted */ }

Transactions memory pool

Methods

impl MemoryPool

pub fn new() -> Self

Creates new memory pool

pub fn insert_verified(&mut self, t: IndexedTransaction)

Insert verified transaction to the MemoryPool

pub fn iter(&self, strategy: OrderingStrategy) -> MemoryPoolIterator

Iterator over memory pool transactions according to specified strategy

pub fn remove_by_hash(&mut self, h: &H256) -> Option<Transaction>

Removes single transaction by its hash. All descedants remain in the pool.

pub fn check_double_spend(
&self,
transaction: &Transaction
) -> DoubleSpendCheckResult

Checks if transaction spends some outputs, already spent by inpool transactions.

pub fn remove_by_prevout(
&mut self,
prevout: &OutPoint
) -> Option<Vec<IndexedTransaction>>

Removes transaction (and all its descendants) which has spent given output

pub fn read_by_hash(&self, h: &H256) -> Option<&Transaction>

Reads single transaction by its hash.

pub fn read_with_strategy(&mut self, strategy: OrderingStrategy) -> Option<H256>

Reads hash of the 'top' transaction from the MemoryPool using selected strategy. Ancestors are always returned before descendant transactions.

pub fn read_n_with_strategy(
&mut self,
n: usize,
strategy: OrderingStrategy
) -> Vec<H256>

Reads hashes of up to n transactions from the MemoryPool, using selected strategy. Ancestors are always returned before descendant transactions. Use this function with care, only if really needed (heavy memory usage)

pub fn remove_with_strategy(
&mut self,
strategy: OrderingStrategy
) -> Option<IndexedTransaction>

Removes the 'top' transaction from the MemoryPool using selected strategy. Ancestors are always removed before descendant transactions.

pub fn remove_n_with_strategy(
&mut self,
n: usize,
strategy: OrderingStrategy
) -> Vec<IndexedTransaction>

Removes up to n transactions from the MemoryPool, using selected strategy. Ancestors are always removed before descendant transactions.

pub fn set_virtual_fee(&mut self, h: &H256, virtual_fee: i64)

Set miner virtual fee for transaction

pub fn get(&self, hash: &H256) -> Option<&Transaction>

Get transaction by hash

pub fn contains(&self, hash: &H256) -> bool

Checks if transaction is in the mempool

pub fn information(&self) -> Information

Returns information on MemoryPool (as in GetMemPoolInfo RPC) https://bitcoin.org/en/developer-reference#getmempoolinfo

pub fn get_transactions_ids(&self) -> Vec<H256>

Returns TXIDs of all transactions in MemoryPool (as in GetRawMemPool RPC) https://bitcoin.org/en/developer-reference#getrawmempool

pub fn is_spent(&self, prevout: &OutPoint) -> bool

Returns true if output was spent