Struct miner::NonFinalDoubleSpendSet

pub struct NonFinalDoubleSpendSet {
    pub double_spends: HashSet<HashedOutPoint>,
    pub dependent_spends: HashSet<HashedOutPoint>,
}

Set of transaction outputs, which can be replaced if newer transaction replaces non-final transaction in memory pool

Fields
double_spends: HashSet<HashedOutPoint>

Double-spend outputs (outputs of newer transaction, which are also spent by nonfinal transactions of mempool)

dependent_spends: HashSet<HashedOutPoint>

Outputs which also will be removed from memory pool in case of newer transaction insertion (i.e. outputs of nonfinal transactions && their descendants)